requirements.txt


Flask>=2.2.0
SQLAlchemy>=1.4.0
jira>=3.0.1
python-dotenv>=0.19.0